Added new features

+++ Application/Locations/Edit.cs
@@ -17,8 +17,8 @@ namespace Application.Locations
             public int Id { get; set; }
             public string Name { get; set; }
             public string Address { get; set; }
-            public float? Longitude { get; set; }
-            public float? Latitude { get; set; }
+            public double? Longitude { get; set; }
+            public double? Latitude { get; set; }
             public int? CityId { get; set; }
         }
